Modified security/roles assignment: base_category can now be used to order categories while building the group id. this way we can assure that the groups will be consistent through all the system (eg. no conflict between 'LIL_NXD' and 'NXD_LIL')
to have the id built on (site, group) when group is fetched from an assignment
and site is statically defined, just put this in the definition:

####
Base Category:
site group

Category:
site/lille
####

Not specifiying 'site' in 'Base Category' will simply append 'site' to base_category
when determining the order and will result in (group, site) instead.
This can cause a "conflict" : if somewhere else, you choose to fetch site and statically
define group, the order will be (site, group), and thus, the two group names won't match.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@3739 20353a03-c40f-0410-a6d1-a30d3c3de9de