Added a variation of Base_edit. This script was created by Romain for solve one edition problem for Planning Box.
The Problem was:

When you have the Planning Box for any Module and consider the user does not have modify permission on the module, it was not be possible to edit the planning box (blocs represent other objects not the module).

If If there is wrong, missing or missunderstood with this commit, please let me know.

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@16921 20353a03-c40f-0410-a6d1-a30d3c3de9de