Added interaction workflow to Packing Lists and orders. The behavior is set resource and quantity at [Order | Packing List] lines when you aggregate items.
To prevent inconvenient edits the resource is only set if the items use the same resource and if resource in the line is empty.
Quantity only changes or are set if the order quantity < items quantity.

Maybe this behavior is inconvenient for some cases, let me know all cases to improve it. Tests      will be added soon.

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@21334 20353a03-c40f-0410-a6d1-a30d3c3de9de