2008-06-16 jerome * use getTitle to show title of current document in <title> in view mode.
title_or_id was used, but it was not following the same semantics as getTitle now uses.
Now that getTitle always returns something, we can use getTitle directly.
Note that this is temporary, the plan is to use getTranslatedTitle in view, list and dialog mode and configure correctly on each portal type how translation of title property should behave.

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@21623 20353a03-c40f-0410-a6d1-a30d3c3de9de