Added new property at RelationFields and MultiRelationField.
Now is possible to define a set of listboxes that will used in dialog mode. This allow the user chooser (if len > 1) which listbox columns/search, he prefer use to find a certain object.

The approach is diferent of relation_form and define columns but still compatible with previous implementation. If the [multi]relation fields are not updated, be previous behaviour will still working.



git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@22688 20353a03-c40f-0410-a6d1-a30d3c3de9de