Fix a dangerous bug happenning in runUnitTest --save :
Since r23751, unit tests instance have a folder for Extensions, and this folder
is cleaned up before --save. Before r23751, this folder was a symlink to the
real instance home's Extensions, so if you run --save on a test instance
initialized before this change, extensions from your real instance will be
removed.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@24206 20353a03-c40f-0410-a6d1-a30d3c3de9de