The Form Printout enables to create a ODF document, using an ERP5 Form and an ODF template.
I and yo had long discussion, whether to invoke
an ODF template or an ERP5 form.
We selected an ERP5 Form to invoke, but because of 
some point of view, we changed the choice and decided
to create a new class to invoke to create a ODF document.

Why we decided to create a new class:

* Reusable of forms 
  If we invoke an ERP5 form to create an ODF document, we have to
  set a ODF template, and the view of the form fixes ODF.
  Thus the form can not use other cases such as html view.
  
* Consistency 
  Set a ODF template as well as page template, the forms has
  two ways two template. It has not consistency.

The Form Printout currently supports only ODT format document.
And the functions only supports paragraphs and tables.



git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@26251 20353a03-c40f-0410-a6d1-a30d3c3de9de