Do not use WebSection_zGetDocumentValueList when portal_catalog is enough.
Simplify WebSection_zGetDocumentValueList to the case it as to handle in order to improve maintainance.

WebSection_zGetDocumentValueList heavily depends on MySQL implementation, as it seems that the "GROUP BY reference" returns the first line of the subselect (which is the expected one due to "ORDER BY priority DESC"). But, without any explicit reference of this behaviour in MySQL documentation, wrong results may be returned with another MySQL version.

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@26745 20353a03-c40f-0410-a6d1-a30d3c3de9de