Remove REQUEST parameter to isHideRowsOnNoSearchCriterion, as it it already available on the ListboxRenderer object, and as it could prevent caching on lazyMethod.
Always display content in Report mode.

Always display content in Domain Mode, when a domain is selected.

If the parameter ignore_hide_rows is found in the REQUEST, always display
content.

Always diplay content when the workflow state parameter is found in the
selection, in order to allow lines to be displayed by the worklists.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@28333 20353a03-c40f-0410-a6d1-a30d3c3de9de