Do not call 'apply' on an existing 'security' attribute once the class is built
More generally, the 'security' attribute should not be used once the class is
built, because the constructor of classes inherited from SimpleItem deletes it.
This is currently tolerated in ERP5 because there is always CopyContainer that
provides one, but it also means that CopyContainer.security is just a mix of
attributes generated automatically on most classes.

Previous code was equivalent to
  CopyContainer.security.declarePrivate('doSomethingStupid')
  CopyContainer.security.apply(Organisation)
which is nonsense.

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@28730 20353a03-c40f-0410-a6d1-a30d3c3de9de