store cache data in the same way for all Image.py methods
Image were not stored in the same way in every methods :
in case of index_html, the image was stored, in case of convert, the image.data
was stored.
This lead to conflicts because in both cases, the cache_id was the same, so if
inded_html was first call, and after that, convert was call on the same image,
convert will return an Image object instead of image.data String.

Now in all cases, the image is stored by setConversion, and in case of convert,
only image.data is returned.

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@31532 20353a03-c40f-0410-a6d1-a30d3c3de9de