Keep text entered by the user.
change a bit the behaviour to keep text entered by the user (if the focus is lost and gain, the text entered by the user is kept, this permit to complete a search after changing the focus).

Put back the version number commited in r31536 by mistake.

reviewed and approved by jm

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@31537 20353a03-c40f-0410-a6d1-a30d3c3de9de