Write "Search" in the bottom right search.
It was reported that in some web_site (like km), some people think that this field was used to login. Writing "Search" inside make not possible to think this now. (And when the field get the focus, the field make him empty to be able to receive the search text).

I try this without javascript support, and the search is usable, the only thing is that the user have to remove the default text manualy (it's the behaviour of most of the popular websites).

reviewed and approved by jm

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@31539 20353a03-c40f-0410-a6d1-a30d3c3de9de