- check alarm_date presence on first and next reindexations
Note: test_21_AlarmCatalogPresence proves that there is a bug in Alarm reindexation.
When new alarm enabled and configured alarm is added to site using one transaction
and one reindexation it will be not properly cataloged.

This affects alarm installation from Business Templates.

This problem was hidden until fix for multiple-reindexation in Business Templates
was applied.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@31572 20353a03-c40f-0410-a6d1-a30d3c3de9de