disable acquisition_copy_value on quantity_unit.
This had been activated in 24427 by Mame, and we do not remember the reason of
this change.
The testInvoice test added in r24428, which looks related to that change, is
passing as of today with acquisition_copy disabled.

On the other hand, the copy behavior is annoying for children of Resource
Documents: Measure, for example, and the soon to come Quantity Unit Definition
Group:
  in those cases, both Product and its children have a quantity_unit, but their
  meanings are different, and the copy is not wanted.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@31722 20353a03-c40f-0410-a6d1-a30d3c3de9de