Add a Manager proxy role.
This fixes cases where a user accesses a bank account of another agency
(this can happen for operation imports at country scale).

It is safe to set a Manager proxy role on this script as it is called on
the only accessed document, so security checks were already done (or
bypassed) upon fetching our context.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@31824 20353a03-c40f-0410-a6d1-a30d3c3de9de