- merge instance and site creation in one step
Work in progress.

Zope and site are controlled by common set of parameters. It is understood that
Zeo will behave slightly different but in case of standalone instance it is
better to not have two different steps to do one thing. Zeo support shall be
plugged between creation of instance and site itself.



git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@32510 20353a03-c40f-0410-a6d1-a30d3c3de9de