- test each category object before expiration
In 32760 isTransitionPossible was used to test if portal type is associated
with a workflow providing such transition. It was wrong oversimplification.

It is required to test each object if expiration can be made, so it will solve
both issues - object is not related with workflow providing such action and
object is in state, from which expiration is not possible.



git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@32764 20353a03-c40f-0410-a6d1-a30d3c3de9de