- create database while creting ERP5 site
MySQL database is *required* part of ERP5 site, so it have to be created.
As all parameters to access mysql with proper privileges are known, and MySQLdb
library might be available for recipe there is no reason to not create database.

Previous method by doing tricks with command recipe is to unreliable and much
less buildout like.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@33086 20353a03-c40f-0410-a6d1-a30d3c3de9de