Fix full-text searches with mixed modes.
When full-text expression uses both boolean and natural mode, the expressions
should be merged, otherwise there would be 2 different definitions for the same
(fulltext) column alias in the SELECT or ORDER BY expressions. If any of the
full-text expressions contain boolean operators, the whole expression will be
executed in boolean mode.

Reviewed by Vincent.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@33656 20353a03-c40f-0410-a6d1-a30d3c3de9de