Fix regression introduced in r33653.
As the generated order-by expressions now (rightly) differ based on data, the
exception generated when detecting the difference must be postponed to when
expression is actualy used, as it might not be used.
Also, test that postponing effectively works, and demonstrate another way to
express the same condition, but which allows sorting.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@33660 20353a03-c40f-0410-a6d1-a30d3c3de9de