Re-ident with 2 spaces (instead 4). Start to handle and monitor openoffice pid like oood ones.
The oood after high usage, deletes the openoffice pidfile and it tries to start a new openoffice while the current one is busy. This makes oood "lost" the openoffice pid file and after been stopped by supervisor the openoffice is not stopped.

This changes make sure the open office will be stopped after shutdown.



git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@35506 20353a03-c40f-0410-a6d1-a30d3c3de9de