fix TestTemplate.test_NonDeliveryTemplateNotIndexable.
Calling directly .unindexObject on a new object is not enough, because
some activity might be queued, waiting to reindex that object. The fact
that isIndexable is False prevents _new_ reindexing activities to be queued
but _not_ already queued activities to go on. In practice, unindexObject
was executed immediatly, a commit/tic followed and executed just after
a reindex, eventually indexing the object.

Queuing the unindexObject call as an SQLQueue activity with an
after_path_and_method_id matching the possible reindexing activities on that
object is the way to go to avoid that reindexation race.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@35553 20353a03-c40f-0410-a6d1-a30d3c3de9de