* Do not invoke simulation_movement_causality_interaction_workflow/interactions/calculate_causality for temporary documents.
Well, calling 'python:not here.isTempDocument()' needs a bit of time. So it becomes a bit slower for non-temporary documents. If we often want to have 'isTempDocument()' check for interactions, adding 'Ignore Temporary Document' option in InteractionDefinition class can be a good idea to reduce the load of preparing python expresison context.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@35995 20353a03-c40f-0410-a6d1-a30d3c3de9de