Optimize isBuildable towards most common case, i.e. case where Business Process is in the same direction than the Simulation Tree flow.
First look up simulation tree ancestors to see if movements related
to business path ancestors are indeed in this ancestor set.
Most of the time this check is enough. Conveniently, it's also cheap.

And if we're unlucky and have a Business Process going upwards, we
go through descendants of the movement (which is costly) to check
buildability.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@36588 20353a03-c40f-0410-a6d1-a30d3c3de9de