do not load in step #3 the objects we've already loaded in step #2
The idea is that if catalog is up-to-date, the call in #2 will already return
and load from ZODB all descendants of self. We do not want in this case to
blindly/dumbly recurse over the descendants of self in step #3, loading them
again from ZODB.

We thus remember which objects we loaded in the Simulation Tree during #2 and
do not load them again in #3. Code complexity is higher, but this should make
step #3 virtually "free" when catalog is up-to-date.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@36761 20353a03-c40f-0410-a6d1-a30d3c3de9de