Move SimulationMovement.isBuildable implementation into BusinessPath.filterBuildableMovementList
The idea was simple: if BusinessPath.filterBuildableMovementList is efficient,
we can replace SimulationMovement.isBuildable by:
  return len(business_path.filterBuildableMovementList([self])) == 1

I thus moved and adapted the code from isBuildable to generalize it
to BusinessPath.filterBuildableMovementList.
Performance remains the very _same_ for isBuildable, but this move allows
us to use filterBuildableMovementList directly in a Global Builder.



git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@37116 20353a03-c40f-0410-a6d1-a30d3c3de9de