One of dependencies of zope.testbrowser is zope.interface.
When zope.interface is installed as eggs (tried with 3.3.0, 3.5.2, 3.6.1) and 
PluginRegistry==1.1.2 is used, there is some bad behaviour which makes erp5 does 
not work. 

It is preferred keep using standard zope.interface provided by Zope-2.8 for now 
until this problem got solved or some good zope.interface version be selected.



git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@37228 20353a03-c40f-0410-a6d1-a30d3c3de9de