Generic jQuery extension to provide autocompletion on fields.
ERP5Site_getCompletionList serves as a backend for jQuery, and
polls the catalog. getCompletionList should be extensible and quite
flexible, even if it only supports portal_type and one key filtering for
now.

Most common usage should be something like:

   $("input.custom_class_defined_in_field").ERP5Autocomplete({
     portal_type: "Product",
     search_key: "title",
   });

 ( produces catalog(portal_type="Product", title="xy%", limit=20) )

but any catalog parameter can used as a key, which means that

   $("input.custom_class_defined_in_field").ERP5Autocomplete({
     portal_type: "Product",
     search_key: "xxcustomer_product_reference",
   });

will work if xxcustomer_product_reference is a custom catalog key



git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@37465 20353a03-c40f-0410-a6d1-a30d3c3de9de