Fix CashMovementNewNotEmitted not really moving containers.
It is incorrect for CashContainer class not to inherit from BaobabMixin, as it
needs "Baobab" accessors.
It is useless for  CashContainer to define manually accessors which are
properly and generically defined in BaobabMixin.
Add missing type-based method so that Cash Containers get a correct
destination, so they actually move containers around (as per Trackinglist
meaning).


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@38108 20353a03-c40f-0410-a6d1-a30d3c3de9de