Speed up installation of Python Scripts and prepare clean up of XML
Before this patch, a Python Script was compiled several times during its
installation. Now, it is compiled only once.

'_code', 'func_code' and 'func_defaults' attributes are also not required
anymore in the XML representation of Python Scripts. This will allow to remove
this garbage for XML in the future.

For the moment, we have to provide forward compatibility with old revision of
ERP5 product, so this patch does not change exported XML.

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@38348 20353a03-c40f-0410-a6d1-a30d3c3de9de