workaround for possible too many open memcached connections problem.
in _initialiseConnection(), try to disconnect existing connections if exist.

python-memcached's get() returns None in case of any error, but it also returns None if None is set, of course. so if 'reconnect' is unfortunately triggered by None value, we should have a living connection at that time.

to fix this issue perfectly, we should rewrite memcached client library that raises an Exception in case of exception.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@39553 20353a03-c40f-0410-a6d1-a30d3c3de9de