fix test_security_defined_on_class after portal type classes changes
This test checks that wrapping a method in an interaction workflow adds
security to the wrapped method, but that it does not override existing
(class) security definition.

To do so, a private method (doSomethingStupid) is added on Organisation
document class. We really want to change the (static) document class here,
not the portal type class: before portal type classes changes,
ERP5Type.Document.Organisation.Organisation was a copy of the document
ERP5.Document.Organisation.Organisation. After portal classes, the former
is a portal type class, and it does not make sense in test logic to add
a method on the portal type class.

It is correct to change the import path to access the physical Document class.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@39652 20353a03-c40f-0410-a6d1-a30d3c3de9de