Add extra method on DCWorkflow _executeMetaTransition to jump from workflow state to another. This will ignore transitions and Scripts.
This meta_transition will be used to migrate documents from portal_type to another.
Because we can not expect that associated workflows will be identical.
The best we can do is pushing the new document to same state of previous
document if exists and this no matter id of workflows.

This feature will be also used by SynchronisationTool to synchronise workflows status.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@40958 20353a03-c40f-0410-a6d1-a30d3c3de9de