- use slapos.rebootstrap in version 2.0 (only 2.12 flavour)
Python will be provided by buildout, but outside of usual buildout environment
(with own parts-directory and installed file). Because of this python binary
which is used to run buildout will no more suffer from uncontrolled rebuilds of
python or its dependencies, as python will be not automatically updated. By
changing rebootstrap:version value it is possible to *add* new python
compilation or switch to existing one.

As, even if python is provided by "same" buildout, but it is out of usual
bin/buildout control, there is no need to recompile it (after rebootstrap is
run buildout:executable becomes "self provided python"). So that is why python
compilation is avoided and python with dependencies can be safely deleted.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@40977 20353a03-c40f-0410-a6d1-a30d3c3de9de