- drop subversion dependency
zc.buildout has feature of extending its configuration over the network.
Thanks to using extends-cache it is possible to have online software
with offline instance -- they just need to be properly directed to
existing extends-cache, and whenever run in offline mode they will try
to reuse configuration files provided there and *fail* (which is good)
in case of missing profiles.

buildout-2.12.cfg extends profiles/official-2.12.cfg only to fill this
extends cache.

Readme is updated to follow the way, thanks to which less dependecies
are required.

As buildout shall work with checkout extends-cache is moved up and
reused from here and generated instance template is generated in place.
This template is renamed to instance.inc, as it is file which shall be
included by instance configuration.

Note: There is no need anymore to use svn to play with ERP5 Appliance
buildout, but it is still possible.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@41035 20353a03-c40f-0410-a6d1-a30d3c3de9de