Use BankAccount_getAvailablePosition to factorise code.
Also, serialise bank account when checking if an operation can happen,
so it doesn't have to be done in each caller script.
This should not be a performance problem to serialise multiple time the
same account per transaction, and BankAccount_checkBalance should
only be called when there is an intent to withdraw cash from the account.

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@41551 20353a03-c40f-0410-a6d1-a30d3c3de9de