Due to the sorting of database connectors during a transaction, the ZMySQLDA connector for the activities database would finish its commit procedure before ZODB, making the description of an activity message in MySQL available before its respective data in the ZODB.
The fix consisted in replacing the ZMySQLDA connector with another
one based on ZMySQLDA but with a “sortKey()” method that forced it
to be sorted after both the ZODB connection and the ZMySQLDA
connection for ZSQLCatalog.

Analysis of issue was done by Sebastien and Julien.

This patch itself was done by Leonardo.

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@41598 20353a03-c40f-0410-a6d1-a30d3c3de9de