MappedValue: remove xxxProperty/edit methods to revert to pre-39918 behaviour
Note that type class of Transformation {Optional,Transformed Resource} Resource
is changed to a document inheriting Amount because they are chained to
conversion_interaction_workflow. Otherwise, _setProperty (that is not redefined
anymore) is would point to _doNothing.
The choice of TransformedResource class is questionable. Alternative would be
AmountFilter, or a new class.

This reapplies r41691

git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@42273 20353a03-c40f-0410-a6d1-a30d3c3de9de