and finally avoid iterating twice over the interaction list
It seems that the only methods that we wanted to match in the second pass
are class methods created during the first pass:
finding the newly created candidates is easy if we keep track of class methods before and after the first pass

And then, the ony interactions than *can* augment the existing workflow definitions
are the ones using regular expressions, according to the code comments.
Queue those interactions during the first pass to reduce the space search.

Finally, the second pass is cheaper because we *know* that methods do exist
and _are_ Workflow methods already.


The overall cost of creating workflow methods should be lower thanks to those
efforts.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@42497 20353a03-c40f-0410-a6d1-a30d3c3de9de