convert testERP5Type to ZODB property sheets
Main guidelines:
* no more property sheet creation through ClassTool,
  use Property Sheet Tool instead
* ban _aq_reset and use resetDynamicDocuments instead
* everytime an action on a portal type/workflow is done,
  we have to wait for the transaction boundary, as those
  workflows delay the resetDynamicDocuments at the end
  of the transaction. hence the many new commits()

This test is too big and poorly isolated.
I expect this commit to add as much as 4 errors and 4 failures
when run sequentially; but when each of those errors/failures
are run individually, they do not fail.
This issue needs more investigation, but is not related to the
migration, hence the early commit.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@42825 20353a03-c40f-0410-a6d1-a30d3c3de9de