refine logging to avoid being pedantic when installing Sites.
I need some way to distinguish what happens when, later, an automatic migration
happens, or when a Tool is not found: if this happens in a "normal" context,
then logging should be very loud and verbose, as this is likely to crash your
site.
On the other hand, during bootstrap/setup, this is perfectly normal to miss
some tools/modules, and we should not worry/log about it, or the NECESSARY
logs will be quickly removed by developers tired of seeing "useless" logs...

Better ideas that can work in a fresh site without workflows, portal types,
or any advanced mechanism? I'm all ears.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@42877 20353a03-c40f-0410-a6d1-a30d3c3de9de