Add a testcase and fix the usage when one adds an empty, or invalid, or incorrect property to a propertysheet.
It was okay, with filesystem property sheets to raise an error: worst case
scenario, the developer would have to fix the property sheet and restart the
instance. But with ZODB property sheets, raising an error can potentially
break the instance. We replace the error by a LOG: you're likely to see this
when editing properties.

This is not a "dirty quickfix", this is not a workaround to go faster: this
behavior will stay. If you feel that it's a wrong way to do, do tell us.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@43199 20353a03-c40f-0410-a6d1-a30d3c3de9de