fix wrong test that expected edit() to call the public setter.
It happened sometime, with aq_dynamic, when a setter was overriden on the class
itself, that setFoo was called on a edit(foo=bar), as class methods were taking
priority over aq_dynamic-generated-setters.  Now, the correct behavior is back:
edit(foo=bar) calls the private setter directly, and never the public setter.

See "Workflow triggers and edit/_edit calling (or not) private or public
setters" email for reference.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@43304 20353a03-c40f-0410-a6d1-a30d3c3de9de