Don't use getattr to access subobjects on an acquisiton-wrapped object.
If some looked-up id is missing, it would be searched in acquisition, which is
not desired.
If some looked-up id exists as a method id on container's class, it will be
fetched instead of a subobject, which is not desired.

Example case, which happened to me: searching for "Title" returns an accessor
from CMF.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@44185 20353a03-c40f-0410-a6d1-a30d3c3de9de