- change behavoiour and do *NOT* call mysql_install unconditionally
Unfortunately mysql_install dies in case if there is software upgrade.
And from current knowledge the only "portable" way to know that mysql
database was initialised is to check for existence of mysql directory in
datadir, which contains system tables.

Such check assumes that installation went fine and then tries to start
mysql. If installation was not correct mysqld will die, which is
acceptable, as wrapper will exit.


git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@44200 20353a03-c40f-0410-a6d1-a30d3c3de9de