on movement collections, use list of properties defined in rules
Until now, we had in movement collections all possible properties
found by _propertyMap for every movement. Now movement collections
use list of properties defined in rules.

This change was breaking some tests because not enough properties
were expanded, so in the same time it is required to add more
properties to progagate on several rules