Do not require empty 'help' directories to be created in Products.
Upon products installation, registerHelp in initializeProduct was first called
with clear=1 for 'help' directory and then clear=0 for 'interfaces' to allow
registering Help for two directories. However, that didn't work if the former
directory does not exist.

Remove such limitation as well as 'help' and 'interfaces' directories which
used to contain dummy files.

Signed-off-by: Julien Muchembled <jm@nexedi.com>
Signed-off-by: Arnaud Fontaine <arnaud.fontaine@nexedi.com>